split the tests up
